[Refactor] Intro SpillCore Module
