+++ be/src/exec/CMakeLists.txt
@@ -34,6 +34,20 @@ ADD_BE_LIB(ExecCore
 
 target_link_libraries(ExecCore PUBLIC ExprCore RuntimeCore)
 
+set(SPILL_CORE_FILES
+    spill/dir_manager.cpp
+    spill/block_reader.cpp
+    spill/log_block_manager.cpp
+    spill/file_block_manager.cpp
+    spill/hybird_block_manager.cpp
+)
+
+ADD_BE_LIB(SpillCore
+    ${SPILL_CORE_FILES}
+)
+
+target_link_libraries(SpillCore PUBLIC FSCore IOCore Common Base Gutil StarRocksGen)
+
 set(PIPELINE_PRIMITIVES_SRCS
     pipeline/operator.cpp
 )
@@ -64,6 +78,8 @@ set(PIPELINE_RUNTIME_SRCS
     pipeline/runtime_filter_types.cpp
     workgroup/pipeline_executor_set.cpp
     workgroup/pipeline_executor_set_manager.cpp
+    spill/operator_mem_resource_manager.cpp
+    spill/query_spill_manager.cpp
 )
 
 set(PIPELINE_ADAPTERS_SRCS
@@ -197,16 +213,9 @@ set(PIPELINE_EXTENSIONS_SRCS
     spill/spiller_factory.cpp
     spill/spill_components.cpp
     spill/mem_table.cpp
-    spill/dir_manager.cpp
     spill/serde.cpp
     spill/input_stream.cpp
     spill/data_stream.cpp
-    spill/block_reader.cpp
-    spill/log_block_manager.cpp
-    spill/file_block_manager.cpp
-    spill/hybird_block_manager.cpp
-    spill/operator_mem_resource_manager.cpp
-    spill/query_spill_manager.cpp
     stream/state/mem_state_table.cpp
     stream/aggregate/agg_state_data.cpp
     stream/aggregate/agg_group_state.cpp
@@ -216,6 +225,7 @@ set(PIPELINE_EXTENSIONS_SRCS
 )
 
 set(EXEC_NON_PIPELINE_FILES
+    spill/spill_dir_manager_bootstrap.cpp
     fetch_node.cpp
     lookup_node.cpp
     capture_version_node.cpp
@@ -451,4 +461,4 @@ ADD_BE_LIB(Exec
     $<TARGET_OBJECTS:join_hash_map_inst_objs>
 )
 
-target_link_libraries(Exec PUBLIC ExecCore)
+target_link_libraries(Exec PUBLIC ExecCore SpillCore)
