+++ be/src/exec/spill/query_spill_manager.h
@@ -14,14 +14,10 @@
 
 #pragma once
 
-#include <atomic>
-#include <mutex>
-#include <string>
-#include <unordered_map>
-#include <vector>
+#include <memory>
 
+#include "exec/spill/global_spill_manager.h"
 #include "exec/spill/log_block_manager.h"
-#include "fs/fs.h"
 #include "gen_cpp/Types_types.h"
 
 namespace starrocks {
@@ -29,22 +25,6 @@ class TQueryOptions;
 }
 
 namespace starrocks::spill {
-class GlobalSpillManager {
-public:
-    size_t spillable_operators() const { return _spillable_operators; }
-
-    void increase_spillable_operators() { _spillable_operators++; }
-    void decrease_spillable_operators() { _spillable_operators--; }
-
-    size_t spill_expected_reserved_bytes() const { return _expected_reserved_bytes; }
-    void inc_reserve_bytes(size_t bytes) { _expected_reserved_bytes += bytes; }
-    void dec_reserve_bytes(size_t bytes) { _expected_reserved_bytes -= bytes; }
-
-private:
-    std::atomic_size_t _spillable_operators = 0;
-    std::atomic_size_t _expected_reserved_bytes = 0;
-};
-
 class QuerySpillManager {
 public:
     QuerySpillManager(const TUniqueId& uid, GlobalSpillManager* global_spill_manager)
